+++ packages/server/src/services/agents/xcm/ops/xcmp.ts
@@ -1,7 +1,7 @@
 import { Observable, bufferCount, filter, map, mergeMap } from 'rxjs'
 
 import { HexString } from '@/lib.js'
-import { filterNonNull } from '@/sdk/index.js'
+import { filterNonNull } from '@/rx/index.js'
 import { createNetworkId } from '@/services/config.js'
 import { ApiContext, BlockEvent } from '@/services/networking/index.js'
 import { NetworkURN } from '@/services/types.js'
