+++ packages/server/src/services/ingress/watcher/head-catcher.ts
@@ -20,7 +20,7 @@ import {
 
 import { BlockInfo } from '@polkadot-api/observable-client'
 
-import { retryWithTruncatedExpBackoff } from '@/sdk/index.js'
+import { retryWithTruncatedExpBackoff } from '@/rx/index.js'
 import { ServiceConfiguration } from '@/services/config.js'
 import { ApiClient, Block } from '@/services/networking/index.js'
 import { BlockNumberRange, ChainHead as ChainTip, HexString } from '@/services/subscriptions/types.js'
